requirements.txt
python-binance==1.0.19
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
ta==0.10.2
ccxt==4.1.13
schedule==1.2.0
loguru==0.7.0
matplotlib==3.7.1
mplfinance==0.12.9b7
plotly==5.18.0
